moved the url builders to its own class, changed some var to let
